# server/docs/swagger.yaml
basePath: /api
definitions:
  internal_handlers.marketplaceManifest:
    properties:
      name:
        type: string
      owner:
        $ref: '#/definitions/internal_handlers.marketplaceOwner'
      plugins:
        items:
          $ref: '#/definitions/internal_handlers.marketplacePluginEntry'
        type: array
    type: object
  internal_handlers.marketplaceOwner:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  internal_handlers.marketplacePluginEntry:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      source:
        $ref: '#/definitions/internal_handlers.marketplacePluginSource'
      strict:
        type: boolean
      tags:
        items:
          type: string
        type: array
      version:
        type: string
    type: object
  internal_handlers.marketplacePluginSource:
    properties:
      path:
        type: string
      ref:
        type: string
      sha:
        type: string
      source:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: CoStrict Team
    url: https://costrict.ai
  description: AI Agent Platform API - Skill marketplace, repository and registry
    management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Costrict Web API
  version: "1.0"
paths:
...
  /marketplace/{repo}/marketplace.json:
    get:
      description: Return installable Git-backed plugins as a standard csc marketplace
        manifest. Plugin sources are pinned to the last successfully synced commit.
      parameters:
      - description: 'Marketplace identity: costrict-plugins, public, or a repository name'
        in: path
        name: repo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.marketplaceManifest'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get marketplace.json
      tags:
      - marketplace
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"